Configure rust-analyzer for Linux target, revert u32 casts

+++ crates/rust-hsm-core/src/pkcs11/keys/cmac.rs
@@ -54,7 +54,7 @@ pub fn gen_cmac_key(
         Attribute::Sign(true),
         Attribute::Verify(true),
         Attribute::Label(key_label.as_bytes().to_vec()),
-        Attribute::ValueLen(cryptoki::types::Ulong::from((bits / 8) as u64)),
+        Attribute::ValueLen(cryptoki::types::Ulong::from((bits / 8) as u32)),
     ];
 
     let mechanism = Mechanism::AesKeyGen;
